[WEBSOCKETS] Adding `wss://` URLs when enabled for Servcies (#714)
~~Since we no longer have a staging environment I have had to relax the
CI so it doesn't panic about not being able to generate the GraphQL
gqlgen which will now need to be done manually. It's not good practice
to enable `introspection` on a production environment. (This now has a
big fat warning in the CI)~~

After [latest changes to
PUB](https://github.com/pokt-foundation/portal-ui-backend/pull/196), CI
is :green_circle: . This PR adds the ability to toggle in the UI based
on the value of `websockets_enabled` retrieved from Portal DB via PUB.

This PR also factors for all changes (and some config changes on Vercel)
related to the removal of the `Staging` Environment. The UI is now
hooked up directly to the PUB/PHD/Portal DB PROD but deploys a preview
environment on branches as they pass the CI.

Updated the [`.env` in
1password](https://start.1password.com/open/i?a=4PU7ZENUCRCRTNSQWQ7PWCV2RM&v=kudw25ob4zcynmzmv2gv4qpkuq&i=c5cretuyeauiubm3uaqojdy4zm&h=my.1password.com)
with the necessary environment variable. (Search: `Grove Portal - Portal
UI - .env (PROD)`)

---------

Co-authored-by: Daniel Olshansky <olshansky.daniel@gmail.com>